fix overflow of web_order_line_item_id in 32-bit devices (#384)
* fixed issues on 32 bit devices where webOrderLineItemId would overflow

* added comment to explain reasoning behind arch check

+++ PurchasesCoreSwiftTests/LocalReceiptParsing/Builders/InAppPurchaseBuilderTests.swift
@@ -21,7 +21,7 @@ class InAppPurchaseBuilderTests: XCTestCase {
     let cancellationDate = Date.from(year: 2019, month: 7, day: 4, hour: 7, minute: 1, second: 45)
     let isInTrialPeriod = false
     let isInIntroOfferPeriod = true
-    let webOrderLineItemId = 897501072
+    let webOrderLineItemId = Int64(897501072)
     let promotionalOfferIdentifier = "com.revenuecat.productPromoOffer"
 
     private let containerFactory = ContainerFactory()
